Atualização Screen Principal
Ajuste da tela principal
